pyproject.toml
[project]
name = "aggregator"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.10"
dependencies = ["python-dotenv>=1.0.1", "requests>=2.32.3"]
